testing many moduli

+++ tools/kernel.c
@@ -25,7 +25,7 @@ void parse_command_line_options(int argc, char **argv)
 	while ((ch = getopt_long(argc, argv, "", longopts, NULL)) != -1) {
 		switch (ch) {
 		case 'p':
-			prime = atoi(optarg);
+			prime = atoll(optarg);
 			break;
 		case 'L':
 			left = 1;
@@ -68,11 +68,11 @@ int main(int argc, char **argv)
 	spasm *U = spasm_echelonize(A, qinv, &opts);
 	spasm_csr_free(A);
 
-        spasm *K = spasm_kernel(U, qinv);
+	spasm *K = spasm_kernel(U, qinv);
 	spasm_save_csr(stdout, K);
 	fprintf(stderr, "Kernel basis matrix is %d x %d with %" PRId64 " nz\n", K->n, K->m, spasm_nnz(K));
 	free(qinv);
 	spasm_csr_free(U);
 	spasm_csr_free(K);
-        exit(EXIT_SUCCESS);
+	exit(EXIT_SUCCESS);
 }
